Change: Clean up text for consistent wording
Clean up text for consistent wording

Dependency: None
